Se crea la validación para Get, Post, Put y Delete

+++ src/routes.js
@@ -1,11 +1,11 @@
 import { Router } from 'express'
-import { PrismaClient } from '@prisma/client'
 import prisma from './lib/prisma.js'
 import validarTanque from './validarTanque.js'
 
 const router = Router();
 
-router.get("/tanques", async (req, res) => {
+router.get("/tanques", async (req, res, next) => {
+    try {
     const { search, tipo, page = 1, limit = 10 } = req.query;
     const tanques = await prisma.tanque.findMany({
         where: {
@@ -25,23 +25,30 @@ router.get("/tanques", async (req, res) => {
     res.status(200).json({
         data: tanques
     });
+    }catch (error) {
+        next(error);
+    }
 });
 
-router.get("/tanques/:id", async (req, res) => {
+router.get("/tanques/:id", async (req, res, next) => {
+    try {
+    const id = Number(req.params.id);
+    if (isNaN(id)) { return res.status(400).json({ error: "ID inválido" });}
     const tanque = await prisma.tanque.findUnique({
-        where: {
-            id: Number(req.params.id)
-        }
+        where: { id }
     });
     if (!tanque) {
-        return res.status(404).json({
-            error: "Tanque no encontrado"
-        });
+        const error = new Error("Tanque no encontrado");
+        error.status = 404;
+        throw error;
     }
     res.status(200).json({
-        data: tanque
+    data: tanque
     });
-});
+    }
+    catch (error) {
+    next(error);
+}});
 
 router.post("/tanques", async (req, res, next) => {
     try {
@@ -59,27 +66,30 @@ router.post("/tanques", async (req, res, next) => {
         });
     }
     catch (error) {
-        console.error(error);
-
-        res.status(500).json({
-        error: "Error interno del servidor"
-    });
+        next(error);
     }
 });
 
-router.put("/tanques/:id", async (req, res) => {
+router.put("/tanques/:id", async (req, res, next) => {
     try{
-        const errores = validarTanque(req.body);
-        if (errores.length > 0) {
-            return res.status(400).json({
-                error: errores
-            });
-        }
-
-        const tanque = await prisma.tanque.update({
-        where: {
-            id: Number(req.params.id)
-        },
+        const id = Number(req.params.id);
+        if (isNaN(id)) { return res.status(400).json({ error: "ID Inexistente"});}
+        const existeTanque = await prisma.tanque.findUnique({
+        where: { id}
+    });
+    if (!existeTanque) {
+        const error = new Error("Tanque no encontrado");
+        error.status = 404;
+        throw error;
+    }
+    const errores = validarTanque(req.body);
+    if (errores.length > 0) {
+        return res.status(400).json({
+            error: errores
+        });
+    }
+    const tanque = await prisma.tanque.update({
+        where: { id },
         data: req.body
     });
     res.status(200).json({
@@ -87,20 +97,30 @@ router.put("/tanques/:id", async (req, res) => {
         });
     }
     catch (error) {
-        console.error(error);
-        res.status(500).json({
-            error: "Error interno del servidor"
-        });
+        next(error);
 }
 });
 
-router.delete("/tanques/:id", async (req, res) => {
+router.delete("/tanques/:id", async (req, res, next) => {
+    try {
+        const id = Number(req.params.id);
+        if (isNaN(id)) { return res.status(400).json({ error: "ID Inexistente" });}
+        const existeTanque = await prisma.tanque.findUnique({
+        where: { id }
+    });
+    if (!existeTanque) {
+        const error = new Error("Tanque no encontrado");
+        error.status = 404;
+        throw error;
+    }
     await prisma.tanque.delete({
-        where: {
-            id: Number(req.params.id)
-        }
+        where: { id }
     });
     res.status(204).send();
+    }
+    catch (error) {
+        next(error);
+    }
 });
 
 export default router
